publish exercise id for cascade deleting (exercise->questions)

+++ CourseServiceAPI/Services/ExerciseService.cs
@@ -3,7 +3,9 @@ using CourseServiceAPI.Interfaces;
 using CourseServiceAPI.Interfaces.Commands;
 using CourseServiceAPI.Interfaces.Queries;
 using CourseServiceAPI.Models.Exercise;
-using CourseServiceAPI.Models.Exercise.DTOs;
+using HackleberrySharedModels.Exceptions;
+using HackleberrySharedModels.Requests;
+using MassTransit;
 
 namespace CourseServiceAPI.Services
 {
@@ -11,13 +13,15 @@ namespace CourseServiceAPI.Services
     {
         private readonly ITableStorageQueryService _tableStorageQueryService;
         private readonly ITableStorageCommandService _tableStorageCommandService;
+        private readonly IPublishEndpoint _publishEndpoint;
         private const string TableName = EntityConstants.ExerciseTableName;
         private const string PartitionKey = EntityConstants.ExercisePartitionKey;
 
-        public ExerciseService(ITableStorageQueryService tableStorageQueryService, ITableStorageCommandService tableStorageCommandService)
+        public ExerciseService(ITableStorageQueryService tableStorageQueryService, ITableStorageCommandService tableStorageCommandService, IPublishEndpoint publishEndpoint)
         {
             _tableStorageQueryService = tableStorageQueryService;
             _tableStorageCommandService = tableStorageCommandService;
+            _publishEndpoint = publishEndpoint;
         }
 
         public async Task<IEnumerable<Exercise>> GetExercisesAsync()
@@ -27,11 +31,21 @@ namespace CourseServiceAPI.Services
 
         public async Task<Exercise> GetExerciseByIdAsync(Guid id)
         {
-            return await _tableStorageQueryService.GetEntityAsync<Exercise>(TableName, PartitionKey, id.ToString());
+            var exercise = await _tableStorageQueryService.GetEntityAsync<Exercise>(TableName, PartitionKey, id.ToString());
+            if (exercise == null)
+            {
+                throw new NotFoundException();
+            }
+            return exercise;
         }
 
         public async Task<Exercise> PutExerciseByIdAsync(Guid id, Exercise exercise)
         {
+            var existingExercise = await _tableStorageQueryService.GetEntityAsync<Exercise>(TableName, PartitionKey, id.ToString());
+            if (existingExercise == null)
+            {
+                throw new NotFoundException();
+            }
             exercise.PartitionKey = PartitionKey;
             exercise.RowKey = id.ToString();
             await _tableStorageCommandService.UpdateEntityAsync(TableName, exercise);
@@ -48,7 +62,19 @@ namespace CourseServiceAPI.Services
 
         public async Task DeleteExerciseAsync(Guid id)
         {
+            var exercise = await _tableStorageQueryService.GetEntityAsync<Exercise>(TableName, PartitionKey, id.ToString());
+            if (exercise == null)
+            {
+                throw new NotFoundException();
+            }
             await _tableStorageQueryService.DeleteEntityAsync(TableName, PartitionKey, id.ToString());
+            await PublishExerciseDeletedEventAsync(id);
+        }
+
+        public async Task PublishExerciseDeletedEventAsync(Guid id)
+        {
+            var exerciseDelete = new ExerciseDeleted { ExerciseId = id };
+            await _publishEndpoint.Publish(exerciseDelete);
         }
     }
 }
