+++ CourseServiceAPI/Services/TopicService.cs
@@ -4,6 +4,7 @@ using CourseServiceAPI.Interfaces.Commands;
 using CourseServiceAPI.Interfaces.Queries;
 using CourseServiceAPI.Models.Exercise;
 using CourseServiceAPI.Models.Topic;
+using HackleberrySharedModels.Exceptions;
 
 namespace CourseServiceAPI.Services
 {
@@ -11,14 +12,17 @@ namespace CourseServiceAPI.Services
     {
         private readonly ITableStorageQueryService _tableStorageQueryService;
         private readonly ITableStorageCommandService _tableStorageCommandService;
+        private readonly IExerciseService _exerciseService;
         private const string TableName = EntityConstants.TopicTableName;
         private const string PartitionKey = EntityConstants.TopicPartitionKey;
 
         public TopicService(ITableStorageQueryService tableStorageQueryService,
-            ITableStorageCommandService tableStorageCommandService)
+            ITableStorageCommandService tableStorageCommandService,
+            IExerciseService exerciseService)
         {
             _tableStorageQueryService = tableStorageQueryService;
             _tableStorageCommandService = tableStorageCommandService;
+            _exerciseService = exerciseService;
         }
 
         public async Task<IEnumerable<Topic>> GetTopicsAsync()
@@ -48,6 +52,10 @@ namespace CourseServiceAPI.Services
         public async Task<Topic> GetTopicByIdAsync(Guid id)
         {
             var topic = await _tableStorageQueryService.GetEntityAsync<Topic>(TableName, PartitionKey, id.ToString());
+            if (topic == null)
+            {
+                throw new NotFoundException();
+            }
 
             var filter = id.ToFilter<Exercise>("TopicId");
             var exercises =
@@ -61,6 +69,12 @@ namespace CourseServiceAPI.Services
 
         public async Task<Topic> PutTopicByIdAsync(Guid id, Topic topic)
         {
+            var existingTopic = await _tableStorageQueryService.GetEntityAsync<Topic>(TableName, PartitionKey, id.ToString());
+            if (existingTopic == null)
+            {
+                throw new NotFoundException();
+            }
+
             topic.PartitionKey = PartitionKey;
             topic.RowKey = id.ToString();
             await _tableStorageCommandService.UpdateEntityAsync(TableName, topic);
@@ -69,15 +83,37 @@ namespace CourseServiceAPI.Services
 
         public async Task DeleteTopicAsync(Guid id)
         {
+            var topic = await _tableStorageQueryService.GetEntityAsync<Topic>(TableName, PartitionKey, id.ToString());
+            if (topic == null)
+            {
+                throw new NotFoundException();
+            }
+
             var filter = id.ToFilter<Exercise>("TopicId");
             var exercises = await _tableStorageQueryService.GetEntitiesByFilterAsync<Exercise>(EntityConstants.ExerciseTableName, filter);
 
-            foreach (var exercise in exercises)
-            {
-                await _tableStorageQueryService.DeleteEntityAsync(EntityConstants.ExerciseTableName, exercise.PartitionKey, exercise.RowKey);
-            }
+            var deletedExercises = new List<Exercise>();
 
-            await _tableStorageQueryService.DeleteEntityAsync(TableName, PartitionKey, id.ToString());
+            try
+            {
+                foreach (var exercise in exercises)
+                {
+                    await _exerciseService.DeleteExerciseAsync(Guid.Parse(exercise.RowKey));
+                    deletedExercises.Add(exercise);
+                }
+
+                await _tableStorageQueryService.DeleteEntityAsync(TableName, PartitionKey, id.ToString());
+            }
+            catch (Exception ex)
+            {
+                foreach (var exercise in deletedExercises)
+                {
+                    await _tableStorageCommandService.AddEntityAsync(EntityConstants.ExerciseTableName, exercise);
+                }
+
+                throw new InternalErrorException();
+            }
         }
+
     }
 }
